adds new UserDetailHeaderCard
