+++ packages/gp2-components/src/index.ts
@@ -30,6 +30,7 @@ export {
   Resources,
   UserCard,
   UserDetailHeader,
+  UserDetailHeaderCard,
   UserNavigation,
   UsersPageBody,
   WorkingGroupCard,
